hotspot2: parse OSU Providers ANQP element
Instead of storing OSU providers list as raw bytes,
parse it using the HSOsuProvidersElement.
Also do not broadcast the raw bytes to the app, since it
will be parsed by us now.

The app will be able to retrieve the parsed information
via a new API in the future (to be implemented).

Bug: 38351209
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: manual testing with Boingo Passpoint AP, verify
      via debug logs

Change-Id: Idc15c4b00580d65133bde57ba43be590d50321c0