hotspot2: track number of providers that connected successfully
Add boolean field "hasEverConnected" to PasspointProvider to indicate
if the given provider had ever provided a successful network connection.
The added field is persisted to the config file.

Update the associated metric for the given provider after it was used
to successfully connect for the first time.

Note that PasspointManager#onPasspointNetworkConnected can be expanded
in the future to keep track of connected SSIDs for a given Passpoint
provider, as an additional input to the network selection algorithm.

Bug: 36358031
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: manually verify the provider's "hasEverConnected" is set to true
      in the config file after a success connection

Change-Id: I6ef4ec801f78f34ff2ca9df97ab06188e1cc0324