SoftApManager: support hidden ap configs
WifiConfigurations for hostapd may set the hiddenSSID flag to true,
intending to start softap without advertising the network.  While this
is not a valid option through the UI, it can be very useful in a test
environment.  This CL adds support to start hostapd with the hidden
network flag set.

Note: an integration test will be added to verify the behavior, but this
requires additional plumbing in the test code.  Tracking that work with
Bug 63141761.

Bug: 34209778
Test: ran modified integration test to verify the ssid is not broadcast
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: If5fef73bc069f155572257e4a9c81efd00dccacf