WifiConfigManager: Method to remove all transient networks
Add a new method to remove all transient (ephemeral or passpoint)
configured networks after every disconect. This is a lazy handling of
the removal of any lingering ephemeral networks in the database. For
example, if we encounter an assoc-reject while trying to connect to an
ephemeral network, this config will now be removed in the next
disconnect.

The alternative is to check and remove the ephemeral network whenever
the connection attempt fails, or if it's ignored after addition because
of the connection rate limiter in WifiConnectivityManager, etc. We chose
the passive removal implementation to balance the tradeoff of fixing the
issue and CL size.

Bug: 62841520
Test: Unit tests
Test: Manual tests with WFA.
Test: Regression tests (b/63663462).
Change-Id: I708955bf6ff59a757b2d4ef7dfbb02b383d8e4d8