WifiVendorHal: Add support for new SAR related HAL API's
Adding support in WifiVendorHal to invoke the 2 new HAL API's added in
IWifiChip in the 1.1 package. The framework needs to be backward
compatible with 1.0 HAL service as well, so these new wrapper methods
return failure if the device is running the 1.0 HAL.
Also, added the plumbing for the new feature bit for this feature:
WifiManager.WIFI_FEATURE_TX_POWER_LIMIT.

TODO: Integrate this with WifiStateMachine in a separate CL which
listens to voice call progress intents.

Bug: 62437848
Test: Unit tests
Change-Id: I9d2b5ba297a5b757933db773eb1b9f067ad80e15