WifiStateMachine: Set Tx power limits during calls
Listen to telephony's call state change events so that we can set/reset
the tx power level to meet SAR requirements in STA mode.
a) When the call is in progress (OFFHOOK), set the tx power limit to
specified level.
b) When the call is done (IDLE), reset the tx power limit (will fallback
to default).

The set/reset is sent to the driver using the 2 new HIDL API's added to
the 1.1 wifi HAL package.

Also, this is an optional feature and we don't need to this unless the
|config_wifi_framework_enable_sar_tx_power_limit| is set on the device.

Bug: 62437848
Test: Manual tests.
Test: Unit tests to be added.
Change-Id: I990f867ce045dfc5f18833afdc9168ec10d929fa