Display notification when wrong password error is detected
A new notification will be generated and displayed when the wrong
password error is detected for the current connection.
It will be dismissed when:
- the user manually dismiss it or tap on it
- attempting a new Wi-Fi network connection
- timeout, currently set to 5 minutes

The current landing page for the notification is the network list page,
it will be updated when the new landing page is ready (dialog prompting
for the new password).

Bug: 33245941
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: manual testing
Change-Id: Iaa75bff6d96ed8c9e9d8472a840d03c54f9f2c52