Move "metered" persistence to WifiConfiguration.
For a long time we've had a nasty tangled dependency between Wi-Fi
and NPMS, since they both persisted different details for configured
networks.  As part of preparing for new carrier data plan APIs, move
the tracking of meteredness over to WifiConfiguration.

Persist both meteredHint and meteredOverride, and flesh out more
documentation for them both.  Allow them to be mutated to support
both upgrade from NPMS and modification by Settings.

When we notice a strong signal that a network is metered, persist
that fact in meteredHint for future reference to be as conservative
as possible; user always has the option to override.

More explicit setting of values in updateCapabilities().

Bug: 63391323
Test: builds, boots, Wi-Fi policy is upgraded
Change-Id: Id7c282b007a23892ee739bd2132c3b63243bd0ce