[AWARE] Create a locally maintained peer ID instead of firmware ID
The firmware provides a peer requestor ID + peer MAC to identify the
peer. The peer requestor ID alone is insufficient.

Original code used the peer requestor ID as the only identifier. This
failed once started testing with >2 devices.

Allocate a local (unique) peer ID which is mapped to the peer requestor
ID and peer MAC by the framework.

Bug: 63642354
Test: all unit tests + integration (sl4a) tests
Change-Id: Ic1b3a97b9675be8d8561f4500e8c777ddff42950