[AWARE] Allow a previously enabled Aware
Current behavior assumes that the firmware is always in sync with
the host - thus the host knows when Aware is disabled on the firmware
and does not expect an 'aware already enabled' error code when trying
to enable Aware.

However, this may not always be the case (e.g. firmware may not disable
aware on wifi getting disabled). Change:

- Considers 'aware already enabled' to be a success
- But prints out a WTF to track (need to make sure that firmware does
  disable Aware when expected)

Bug: 63773428
Test: integrated (acts/sl4a) AttachTest:test_attach_disable_wifi_attach_again
Test: unit tests pass
Change-Id: I940a2ff3e5c24434a1bb2bdf21263b2f8e9769a0