Set nfc.initialized system property after NFC is initialized.
This property will be used by esed as an indication when the deamon can
start. NFC firmware download requires the ese gpio pin low, where as ese
transactions need it to be high. In order to avoid race conditions and
to make sure initialization doesn't fail for either NFC or ESE, NFC
will finish initialization and then set the property. ESE will wait for
this property to be set before starting.

Test: adb shell getprop
Bug: 64010793
Change-Id: I74550efbbaccb2ba252c7b11649f468c60a3ce35