[WifiScoreReport] Include scores in dumpsys
To help diagnose problems in connected-network scoring, include the
computed scores (currently legacy and aggressive) in the output
of dumpsys wifi.

While here, discard any samples that might happen before the clock
is set on a newly flashed device, to avoid confusion due to the
fluffy timestamps.

Also bump the session number only once in the case of back-to-back
reset() calls, since these are quite common.

Bug: 63862121
Test: wifi unit tests
Test: adb shell dumpsys wifi WifiScoreReport
Change-Id: I299a04cd0cf35d81cc179ae9fee6760daf3d8a59