RELAND: HalDeviceManager: Get IWifi proxy immediately
Try to get IWifi proxy as a part of HalDeviceManager.initialize()
directly instead of waiting for the IServiceMAnager callback.

When HalDeviceManager.initialize() is invoked:
a) If the wifi service is up at that point, we will get the proxy immediately
and IServiceManager callback will be ignored.
b) If the wifi service is not up at that point, we will get the proxy as a part
of the IServiceManager callback.

Bug: 64226605
Test: Unit tests
Test: Regression test request.
Change-Id: I3d6b89edb3f446bb39984d1ebb3caeb53011bec9