WifiConfigManager: Skip data migration if new store is present
If the new config store files are present, skip data migration from the
older store file. The old store files may still be present in the device
if WifiConfigStoreLegacy.removeStores() failed for some reason during
the earlier migration.
Also, ensure that we attempt to delete all the store files in
WifiConfigStoreLegacy.removeStores(). Returning early on failure in that
method may leave the device in an incoherent state (some old store files
are deleted, while others are not).

Bug: 63376504
Test: Unit tests
Change-Id: If2e65a7e03677040dfccf88298887e21201aff92