Meteredness should reflect user opinion.
While cleaning up the original meteredness refactoring CL, I must
have dropped the usage of WifiConfiguration.isMetered() which blends
together all the various options to decide if a network should be
treated as metered or not.  Fix that bug.

Also update the capabilities whenever a configuration is updated,
since the user (or tests) may have just changed their meteredness
opinion.

Test: cts-tradefed run commandAndExit cts-dev -m CtsHostsideNetworkTests
Bug: 64274313
Change-Id: I6b4df07b6222810f7b3790a5c2f1685dfe3535c9