Check roaming network for Fast Re-Authentication
Sometimes fast re-authentication doesn't work when the user moves from
the outside into the range of AP because BSSID mismatch is caused.
When the user taps AP for first connection, network block sets BSSID
with "any". And when the user moves into the range of AP,
SavedNetworkEvaluator sets BSSID from scan result.
And current code does not care about roaming case. When device tries
to connect to different APs in the same ESS, Wi-Fi framework always
handles them as the same network and won't update the target BSSID to
supplicant. So supplicant always starts the connection with unexpected
BSSID.

The patch sets BSSID to supplicant if roaming network is detected.
It classifies roaming network if BSSID is different from first
connection.

Bug: 64043254
Test: Unit tests

Change-Id: I4a57e7c05aaee99781dd38ef0405fcfaffae10c8