Remove hostapd.conf before starting Tethering
It fails to start Wi-Fi Tethering at first boot
after OS version up from Android N to O.

It is caused by permission.
Owner of hostapd.conf is "system" on Android N.
However, it is "wifi" on Android O.

hostapd.conf doesn't need to be handed over
because it is created again when Tethering is started.

Bug: 67351973
Change-Id: Ia7c87ab1f2040434963f6e0670ccd02cdd84bee7