WifiConfigManager: Remove permission checks for connect req
When WifiConfigManager was initially written, we wanted to prevent 3rd
party apps from forcing the device to connect to a particular network
using enableNetwork(). But, since we did not have an alternate API
surface exposed to apps in O, we had to temporarily relax this permission check
and let apps connect to networks with a warning log message. Get rid of this
permission check altogether because the log message is misleading since
we don't have any alternatives planned for now.

WifiConfiguration.lastConnectUid is a field used only for debugging
purposes. This was not previously getting updated for connect attempts
by apps with insufficient permission.

Bug: 65703184
Test: Unit tests
Change-Id: I35560cefedeff471128cf54587fa9542e3493066