[RTT][HDM] Remove bound interface from RTT controller creation
When creating the RTT controller, the HAL API allows an optional
interface specification which is the 'bound interface' - i.e. the
interface on which the RTT operations will be run. A null is allowed
which indicates that the HAL will decide where to run the RTT operation.

The null option should be the default mechanism - the RTT operation is
run based on the MAC address of the peer and the firmware can/should
decide where and when and how to run it (e.g. from a STA to an AP or
on NAN).

Bug: 65014872
Test: unit tests and RTT integration tests
Change-Id: I117300ad5a04387e843eaa66e84fc7d04fcf528b