Rework the PacketKeepalive parameters to Separate L2
Rework the packet keepalive parameters as part of the effort
to support NATT keepalive on cellular, which does not have an
Ethernet L2. At the same time, there is some small cleanup
related to how parameters are passed, so the KeepalivePacketData
is discarded earlier in the process, which makes it easier to
follow which parameters are being used where/when.

Bug: 69063212
Test: verified using SL4A b/72420081 tracking unit tests
Change-Id: I4a170d680b88d84e400fdad5bda4351c824b3fea