[AWARE] Restructure debugging flags to enable dynamic updates
Restructure the debugging flags to allow dynamic updates of the debug
flag. Each class has a static VDBG and a dynamically updated mDbg.

Note: there are no (or minimal changes) to the actual debugging - all
of which is gated by VDBG. A subsequent CL will move some messages to
the lower-level mDbg.

Bug: 69432628
Test: builds, runs, toggle debug flag
Change-Id: I61df2a0bb790832c7136aa1ad67a745eb388b1da