WifiNative: Support devices with no vendor HAL
For devices which do not support the vendor HAL, we'll not support any
concurrency. So, simulate the WifiVendorHal interactions to bring down
any existing interfaces when we create a new interface.

Modified the unit tests to use a global InOrder. Creating local inorders
causes inconsistent behaviour across helper methods. Needed this
refactor to help test the changes in this CL.

Bug: 70521011
Test: Unit tests
Change-Id: I44a9dec92570cfa1ad49e297e6b7c8dc103ffeaf