WifiConfigManager: Prevent store writes before a read is triggered
Overwriting the store file before reading from it should not be allowed.
Usually the BOOT_COMPLETED notification is received pretty early on and
that would have triggered the store read before anything else happens.
But, looks like this is not the case in fugu/elfin. So, add checks in
place to reject such write requests.

Also, added a missing @Test annotation on one of the existing unit tests.

Bug: 71504864
Test: Unit tests
Test: Manually verified that the CL fixed the issue reported.
Change-Id: Ib0cf39fd0790c90287cc7219aba46f4e855815e0