WifiConfigStore: Inform the clients when store has no data for them
Some modules like ONA, WakeupController are dependent on
store read completion to kick off operations. So, ensure
that the config store informs all the clients when the
shared/user store is read even if there is nothing in the file (fresh
install).

Changes in the CL:
a) Only invoke deserializeData & resetData for user store if it has
become visible (after user unlock).
b) When the store contents are empty, inform the clients that there is
no data to read by passing in a null parser handle.

With these changes, the clients can effectively wait for the
deserializeData() of their store data module being invoked to
kick off operations.

Bug: 71644289
Test: Unit tests
Test: Manual tests (Added logs in all the clients DataSource to ensure
they all received a call to deserializeData() twice on bootup (one for
shared & one for user).
|in| being null when they have no data & |in| being non-null when they have data)
Test: Sent for regression tests: b/71712638
Change-Id: If250252baf147bb16990dfa693417ead312b9911