[AWARE] Remove ability to accept connections from ANYONE
Wi-Fi Aware data-path (NDP) setup requires an Initiator and a Responder.
Both Initiator and a Responder are set up with the MAC address of the
peer to which to connect to as well as a security configuration (Open,
PMK, Passphrase).

The original API (27) allowed a Responder to be configured to accept
a connection from anyone (e.g. using a null MAC address). This creates
ambiguous semantics when an NDP already exists. The second Responder
request could refer to the previously setup NDP or to a new one to be
created. We cannot tell the difference up-front.

Remove the "Accept request from ANYONE" API for newer APIs.

The service implementation retains the original code allowing (some API
levels) to specify the option.

Bug: 72175022
Test: unit tests and integration tests
Change-Id: Idc0ec67bb6481827340439e77463f6579a994104