WifiScanningService: Reload WifiScannerImpl on every driver load
Changes in the CL:
a) Recreate the scanner impl on every driver load so that we fetch the
interface name dynamically on toggle.
b) If the interface name is null, stop all state machines by returning a
null scanner impl.

TODO: We can eventually use the iface name in the scanner impl's for any
native calls.

Bug: 72431133
Test: Unit tests
Test: Toggled wifi/softap/airplane multiple times.
Test: Rebooted with wifi on/off & ensured that we can connect to wifi
networks.
Change-Id: Ia0ef83371e92815a33db8b9e5b95407118410fbe