WifiStateMachine: remove CMD_NO_NETWORKS_PERIODIC_SCAN
Since scans are triggered by other means, we can now remove the logic to
trigger scans in WSM in the situation where there are not any saved
networks.  This was previously used to make sure the OpenNetwork
Notification could still function.  At this point, scans are triggered
by WifiConnectivityManager.

Bug: 31598147
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: manually verified the ONN popped up without any saved networks
(indicating that a scan was performed)

Change-Id: I5020c32581c1b88642aedf8cc2595f2259d0b1a0