Validate restore of old wifi settings
We want to validate that older backups of wifi settings using KEY_WIFI_SUPPLICANT
and KEY_WIFI_CONFIG (which were removed in O) can be restored without
throwing exceptions on invalid backup data. In this case, we
catch potential null cases and number format exceptions that occur when
parsing the network data.

We also add GTS tests to validate that we don't crash when restoring
these malformed backup data sets and that our checks persist.

Bug: 74060048
Test: 1) gts-tradefed run gts -m GtsBackupHostTestCases -t com.google.android.gts.backup.OldWifiSettingsRestoreHostSideTest
2) frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: I44a3c51b11019a0b292eea6e3a8e4ec417aea5fc