WiFi: Add stub for wait_for_driver_ready()
This commit is part of the change to add a vendor hal function to wait
for the wifi driver ready so frameword does not prematurely start to
access it before it is ready for operation.
This commit adds the stub for the function for platforms that will not
implement this function.

Bug: 73482286
Test: Manual Test
Change-Id: I5fe646f78cce098e0fc5fd10f20faab06688ac05
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>