WiFi: Move SAR implementation to SarManager
This commit is the first stage of extending SAR Support in WiFi.
This stage include removing SAR handling from the WifiStateMachine and
putting it in a separate SarManager.
No new functionality is added yet in this stage.

Note that unit test cases for SAR related functionality in WifiStateMachine
are removed and added to SarManager

Bug: 65174506
Test: Run Wifi unit test suite
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: I0a08180cfff22002f75a6966009e62b9c66044c3
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>