Fix WifiWake locking behavior.
To prevent a single non-representative scan from determining the state
of the WakeupLock, we no longer initialize the lock with solely the
results from WifiScanner.getSingleScanResults(). Instead, we
iteratively build the lock up over 3 scans (capped by a timeout).

The WakeupLock is now responsible for determining whether to add or
remove the networks from the lock based on its initialization state
(see WakeupLock#update()).

All DFS channels are excluded from consideration for both locking and
waking.

The initialization state is not persisted to the StoreData, but rather
any loaded StoreData is assumed to be initialized.

Bug: 74259229
Test: runtest, manual
Change-Id: I062f000546079dee1be1aa565a515237ab0c745f