Send "02:00:00:00:00:00" as device MAC when not connected
When Connected MAC Randomization is on and the device is disconnected,
we don't have a valid MAC address to show since we don't know what the
next MAC address is going to be. So instead of showing the MAC address
used during the previous connection, WifiInfo should return the default
MAC address "02:00:00:00:00:00".

In this change, the device resets MAC address stored in WifiInfo to the
default MAC address whenever it gets disconnected with Connected MAC
Randomization on.

Bug: 74436251
Test: Unittest
Change-Id: Ibc1027559a9d5abead2a764b5e94672498d4f7b4