WifiConfigManager: Validate bitsets in incoming Wificonfiguration
Disallow network additions with invalid bitset values set in the various
bitsets in WifiConfiguration. This may otherwise overwrite any valid
WifiConfiguration objects.

Note: This is a defensive check to guard against malformed
WifiConfiguration objects being passed in to the wifi stack. Does not
fix the root cause of the linked bug. That is still being debugged.

Bug: 73987207
Test: Unit tests
Test: Ran ACTS WifiManagerTest locally.
Change-Id: I6187f1b02951cc0ec11adfdb7b2acccda123d361