WifiNative: Change logging for interface up/down handling
When |setupInterface| method is invoked, we invoke
|onInterfaceStateChanged| helper method to store the state of the
interface as we're exiting the call. So, the first interface up state is
not being logged (although callback is correctly being dispatched to WSM
or SoftApManager). Move the logging to the helper method to ensure that
we consistently log all relevant interface up/down messages.

Bug: 75991970
Test: Verified that the interface up message is being logged when wifi
is turned on.

Change-Id: Ie6e51d4d249c6073b72a256b4ec8b0ad439628e0