WifiPermissionsUtil: update canAccessScanResults
For apps with location and (change or access_ wifi state permissions,
access to scan results (and other sensitive calls) will be granted
when Location is toggled off, but wifi is enabled.  This allows
connectivity apps to continue working even if location services
are off.

Also added the SmallTest annotation to the WifiPermissionsUtilTest while
here.

Bug: 69976988
Test: frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Test: installed wifi analyzer app and confirmed it has access after CL
when location is toggled off (and it didn't before the CL)
Change-Id: Ie17f19aff6096847da5fb2d37b0aa8cd7932f3f4