[WLRW] don't take bugreport if reconnect on new network
After watchdog triggers a restart, if wifi connects to a new network that
the watchdog hasn't been tracking, we should not trigger a bugreport
because what's probably happened is the user moved to a new location.

Bug: 77712034
Test: compile, unit test
Manual test:
[done on a hacked version that made watchdog trigger on 1 failure]
Enter wrong password for a wpa network.
Observe watchdog restarts wifi.
Connect to a open network.
Observe that watchdog does not trigger bugreport.

Change-Id: I39e120e4103458f7e4a82f4fa3a3cdbd9489f9ee