[WIFI][METRICS] Update Passpoint and RTT metrics collection criteria
The Wi-Fi Metrics module collects information on observed APs which
support Passpoint (R1 and R2) and IEEE 802.11mc (RTT). This metrics
collection was added to code which gated metrics collection only for
APs with strong signal.

Update the collection point to always collect metrics for passpoint
and RTT - irrespective of AP signal strength.

Bug: 76416162
Test: runtests.sh -e class com.android.server.wifi.WifiMetricsTest
Change-Id: I29d69218efd9f82e8d42067b93bf3c6910699c74