[AWARE] Failure to release network requests can lead to further failures
Fix failure observed during stress tests: if a data-path is terminated
due to remote operation (one side always finishes first) then it may
causes cascaded failures.

The behavior is due to a failure to clear out the cache from previous
requests. The framework checks to see whether a request can be released,
it should not be released if:

- An agent exists, 'unwanted' will be called on the agent
- Or an NDP was created (even if an agent doesn't exist), should first
  terminate the NDP

But otherwise the request should be cleared from the cache - and it
wasn't.

Bug: 73896170
Test: DataPathTest, DataPathStressTest (and observed failure which weren't cascading)
Change-Id: I1caa8998f071426dfb3e218ed113f4bb3f2a8cff