WiFi: SAR Support: Add SarInfo
Current implementation has limited support for SAR in Wifi.
It just includes the support of power backoff when an ongoing
voice call is happening while Wifi is transmitting.

This commit is the first stage to provide more support for SAR
for WiFi TX power backoff when the device is in close proximity
to head or body of user with or without the presence of a voice call.

This stage adds the SarInfo which is a data structure that is used
by SarManager to communicate SAR triggers towards the WifiVendorHal

Bug: 65174506
Test: Run Wifi unit test suite
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh
Change-Id: Ic5f002430db1e92f656ae861e65934241a8590ef
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>