WiFi: SAR Support: Use SarInfo in WifiVendorHal
In a previous commit, the class SarInfo was introduced.
This commit utilizes this class to pass SAR related inputs to
WifiVendorHal to select the SAR scenario.

Note that the functionality introduced in this commit is not
called yet during operation. Hence the legacy code is left in
place until the new functionality is fully utilized.

Bug: 65174506
Test: Run Wifi unit test suite
Test: ./frameworks/opt/net/wifi/tests/wifitests/runtests.sh

Change-Id: Idb00f2c265d64ea4e8fc3a6e2ed7fc9bc08539fb
Signed-off-by: Ahmed ElArabawy <arabawy@google.com>