[WIFI] Make sure only a single callback is registered on IWifiChip
A callback was registered on the Wi-Fi chip on every new interface.
This resulted in multiple registrations and multiple callbacks - aside
from getting multiple calls eventually this can lead to failures as
we reach the limit of registrations.

Note: this CL assumes a single chip. This reduces the code complexity.
TBD on a more generic solution.

Bug: 79876515
Test: unit test
Test: manually change config - verify single callback
Change-Id: I4582ba7631153f465e57adec13ab4397f1c4691e