fs_mgr: overlayfs skip mount points that are more specific
This is an optimization to limit us to an optimal set of overlay
mounts.  If an entry is identical or more specific (eg: /vendor is
less specific than /vendor/firmware_file), then no need to craft an
override for it.

We do the check in the setup and mount_all paths.

Test: confirm on devices with more specific mounts they do not get override
Bug: 109821005
Change-Id: If34775455fc7989903f70bc4f075c948db91ce37

Former-commit-id: 62aeb16e1d89959893d29ba529a00703c5451000