libprocessgroup: Move CgroupSetupCgroups() to libprocessgroup_setup
Only init uses SetupCgroups. This functionality is
moved from libprocessgroup to its own library, and only
init links to it.

Also, merge CgroupSetupCgroups() with CgroupMap::SetupCgroups()
because the former is just an alias of the latter, and
CgroupMap does not belong to libcgrouprc_setup.

Test: boots
Bug: 123664216
Change-Id: I941dc0c415e2b22ae663d43e30dc7a464687325e
Merged-In: I941dc0c415e2b22ae663d43e30dc7a464687325e

Former-commit-id: 17cb66c3c7ca92b60fda0117036304e026abbbab [formerly 6f9ce2e548e4a583392b212ca9017f6debdeebc0]
Former-commit-id: 6aa9e15e74defa625b1a8e7cd2cf65cb61c843db