libcutils: android_get_control_file uses realpath.
If the path to android_get_control_file is a symlink, the final sanity
check will fail that the fd does not have the same path as the given
file. We can't expect callers to readlink() because this would change
the environment key. Instead, try to call realpath on both paths.

Bug: 126233777
Test: lpdump
Change-Id: I0df10d7dbe3e572b8335faad812e5cd80bff1733

Former-commit-id: 6bcbe47ece10af4a5cf41baef55cf21ae8b1b211 [formerly dd6eefca303755264434454fddc0cc627e834577]
Former-commit-id: 4c21462040dc5dc9d7cc71341528fb0d7d4aaa4c