entry_name_utils-inl.h: clear top bit before shifting
Attempting to shift a uint8_t which has the top bit set results in a
value outside of the range allowable for a uint8_t. The process of
converting this value back to a uint8_t triggers ubsan's
implicit-conversion sanitizer.

Ensure the high order bit is stripped before shifting, so that the
shifted value fits into the allowable uint8_t range.

This change is necessary to enable integer sanitization on this code.

Somewhat related to Bug: 122975762
Test: atest ziparchive-tests

Change-Id: I940c8b2828ac88348d80f731274b990e35de6631

Former-commit-id: 28c7f3eea7c575831363881804d2baaa2184b734 [formerly c5da03f448485044835046fc31f9a734060ffb99]
Former-commit-id: 2bd1f0ededfab02c2cbac2eea05e070066a75fd1