libprocessgroup: use libcgrouprc to read cgroup.rc
Use the LL-NDK library to read cgroup.rc.

As a side-effect, the CgroupController class is changed
to a wrapper of ACgroupController* from the library. The
class itself essentially becomes the pointer, and can be
trivially copied. Modify all client code for this change.

Test: builds and boots
Bug: 123664216

Change-Id: I635d9885db62e82f37421f6d20f59c84a6f4cbb2
Merged-In: I635d9885db62e82f37421f6d20f59c84a6f4cbb2

Former-commit-id: b2d210b52a146cbf2faf8f01f41fe3617d3699d8 [formerly 53e0deb7b287df2f59ca765a845bcca1fb45f5e4]
Former-commit-id: f5d14682d5c41d000e61d5f4aa145e4bdd8098fb