adbd: increment writes_submitted_ before submitting writes.
If we fail to submit writes for some reason (e.g. the USB cable was
unplugged), another thread that's waiting on the write mutex can enter
SubmitWrites and attempt to resubmit the writes that we already failed
to submit, leading to a failed assertion of !IoBlock::pending.

Increment writes_submitted_ before actually calling io_submit, so we
skip over these writes and fall through to exit.

Bug: http://b/129134256
Test: manually unplugged a blueline
Change-Id: I2fd0034e45db22c8f637c81039ce686b7aa6a03b

Former-commit-id: b2d6fe7f9c1295b6745114a85d875abc7c7135ab [formerly 63b52ec13b8b612b79d9c6884776b72bff6bebd0]
Former-commit-id: 8e336bbe6b7fad21f764efdf6a73efb46c5feba8