Avoid leaking property values into logs on error
The purpose of having fine grain read/write control over the property
space is to help ensure the confidentiality of data stored in
properties. Leaking property values into the dmesg buffer on errors
exposes the value outside of the access control rules specified by
policy.

(arguably this is also true for the property name, not just the value.
However, property names are exposed in other places now, so the
incentive to fix this is lower. It would also take away a valuable
debugging tool.)

Test: compiles
Change-Id: I4a0634b8b5e4fd2edf718eaf7343940df627366d

Former-commit-id: b8c5651b39f0a61cee8ca028115a06e100404971 [formerly 9ca898fff81b2cd5adeb77cf1795b4fe8ab07f7f]
Former-commit-id: eef5d0521db99b48e5ce364832c7ee71b403bd7e